Extracting the Harmony patches from Main into their own files: TwoWeaponFightingAttackPenaltyOnEventAboutToTriggerPatch
